+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  users or artists
  - Proper replacements for all deprecated playlist endpoints
  (See https://developer.spotify.com/community/news/2018/06/12/changes-to-playlist-uris/ and below)
+- Allow for OAuth 2.0 authorization by instructing the user to open the URL in a browser instead of opening the browser.
 
 ### Deprecated